backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
playwright
openai
pandas
beautifulsoup4
python-dateutil